Move ImageUploader code to src/singletons/
