+++ src/singletons/ImageUploader.hpp
